requirements.txt
spacy==3.5.0
python-docx==0.8.11
